Proyectos IA /Pa09 /Gargantua Devs/requirements.txt
streamlit
tensorflow
pandas
numpy
PyPDF2
Pillow
pytesseract